(GEN-154) Ensure that hclog.Default is configured and used.
This commit ensures that hclog.DefaultOptions is configured, both for
lyra and its embedded plug-ins.

The commit also removes a global `log` variable since initialization
of it causes a premature instantiation of the Default logger which in
turn makes impossible to subsequently configure it. Changing the
DefaultOptions has no effect once the Default logger has been retrieved.