Experimental Terraform bridge provider and workflow
Proof of concept showing a Lyra provider that bridges to the Terraform AWS provider along with an example "attach" workflow.

This code is directly derived from the Pulumi Terraform bridge and makes use of code generation (from https://github.com/lyraproj/bridge) to provide some boilerplate. Ideally we replace this static generation step with a dynamic runtime solution.

The good news is that most of those 72k additions are generated code ...