Add support for lyra link files
Adds ability to express plugins as '.ll' (Lyra Link) files containing
a command and an argument or array of arguments expressed as yaml. Both
the command and the arguments are subjected to environment expansion.

At present, this should be viewed as a quick and dirty fix to enable
lyra to start a nodejs executable with a file.

Relates to #42.