Separate content plugins from main Lyra binary
To date we've embedded our content plugins (aws/azure/etc) into the main Lyra binary. The binary grew over 200MB as a result.

This commit separates the content plugins into their own binaries. In the near term this gives us a much smaller Lyra binary but longer term it's in line with how we expect plugins to be distributed. All binaries are now also stripped of debug info by default, though panic() stack traces will remain human-friendly. Core plugins related to core functionality - identity and puppetdsl - remain embedded.

One issue this change raised is that the loader needs to pull files from two different locations: the 'plugins' dir as usual which contains workflows and types, but now also the 'build' dir which contains the content plugins. The loader has been modified to load plugins/DSL/YAML/links from a path now rather than just a single directory. That path defaults to "plugins;build". That means for each type of thing we load (DSL or plugin or YAML or link) we check what we can find in the "plugins" dir then check what we can find in the "build" dir. Note that we still load plugins first from all dirs on the path, then DSL from all dirs on the path, then YAML from all dirs etc.

A last interesting note: this commit removes an unused function from "apply.go" which was the sole usage of a dependency called "enry". That was our largest dependency and without it the binary is 7MB smaller...