I18n updates including strings embedded in the binary
This commit slightly changes our approach to i18n by using the English string as message key. The gotext library will return the key if no translation is found and the examples in the docs suggest that this is the intended usage. There's some debate around using abstract keys vs using primary language text. In our case we want to embed at least English in the binary and this apporach gives us a simple way to achieve that. We are i18n-ready but can avoid dealing with shippibng/deploying/finding locale files until we actually support more than one language. It avoids problems with missing translations resulting in abstract keys appearing in the UI. It also makes it clear when retranslation needs to occur, in the case where the message content changes.

We also use gotext directly so that we might be able to employ tooling like this to extract strings: https://github.com/leonelquinteros/gotext/tree/master/cli/xgotext

There are some small tweaks to details:
- actual phrasing e.g. saying "workflow" instead of "activity
- all usage commands have meaningful placeholders
- examples that are the same as usage are removed
- validation is disabled as it doesn't work at all