Switch to golangci-lint
By default, we lint only pkg/... and cmd/lyra/... to keep memory consumption low enough to run on Travis.

Added a "make lint-all" target that runs on all packages.

The build breaks if lint is found and this commit also fixes the handful of lint errors found by default (but not those found by lint-all).