Fast default "make" target but thorough Travis build
This commit mostly updates the Makefile to create one fast and one thorough version of our build.

The default "make" target is now a fast variant of our existing lint/test/build approach. It runs the fast version of the linter, runs tests without race checks and builds binaries without stripping or compressing them. This gives a quick default build but with decent test/lint coverage. This is the experience that a new user who runs a simple "make" will see.

The "all" make target performs the same lint/test/build steps but more thoroughly and hence quite a bit more slowly (approx 5 mins vs 1 min). The full set of lint checks are performed, the tests are run with race detection and the binaries are shrunk/compressed.

Travis will now perform a "make all" to ensure thorough checks on PRs.

The Makefile is also pruned to remove proto and licence-relate things we don't use any more.