Load on demand and use of new directory layout.
This commit replaces the Lyra specific loader with FileBased loaders
capable of loading and starting plugin services, load manifests (yaml
and pp) and lyra links. All PreLoading is removed except the one that
takes place when generating types for all known plugins.

The expected layout of where things are found is also changed. The
former "plugins" directory is now called "workflows" and the "types"
directory that used do reside under "plugins" is now a top level
directory.

The puppet-dsl no longer has a main function in Lyra, instead it is
built directly from the Go module.

The go-plugins no longer has a goplugin prefix to their names. Instead
they end up under the directory "build/goplugins" and have names that
exactly correspond to the TypeSet that they produce.

Closes #80
Relates to #41
Relates to #172