Enable stacktrace for issue.Reported when running with --debug
This commit leverages a new addition to the "issue" package. By
calling issue.IncludeStacktrace() with a "true" argument, all
issue.Reported errors will be created with a stacktrace from
the point they were created. As a result, almost all errors
generated by the framework will be displayed with a stacktrace.

Closes #184