Change Lyra to use the latest terraform bridge.
This commit updates the go.mod to bring in the updated terraform-bridge.
Since that update changed a lot of names, most examples under the
workflows directly are now broken. The only exception is the aws_pp
which was used for testing and therefore updated.

All TypeScript examples were updated since this was necessary to get a
green build.