Switch to Hiera with full configuration capabilities
This commit removes the option to explicitly name a data file for
Hiera data in favor of using a Hiera configuration stored on disk
in the file 'hiera.yaml'. If no such file exists, Lyra will fall back
to Hiera default behavior and first look for a file called 'data.yaml'
and then for 'data/common.yaml'.