Update foobernetes examples to use 'policy' parameter
While debugging hiera lookups, I noticed that although
we had the load balancer policy as the single external
parameter to the workflows, it was neither defined in
the resource type nor used in the body of the workflows.

This commit adds it to the Foobernetes::Loadbalancer
struct, the test data, and the resource definitions
in the workflows themselves.

(NB foober_iter is not fixed in this commit, but it
no longer applies cleanly and I'm not sure we need to
keep it at all)